Split CD-ROM backends into their own modules
